pyproject.toml
[tool.poetry]
name = "mock-ws-server"
version = "0.1.0"
description = ""
authors = ["dianhsu"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
websockets = "^13.0.1"
python-dotenv = "^1.0.1"
loguru = "^0.7.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
